% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctgr_acctg_flds.r
\name{ctgr_acctg_flds}
\alias{ctgr_acctg_flds}
\title{Auto categorize format types to pass into 'fmt_acctg_tbl'}
\usage{
ctgr_acctg_flds(data_tbl, ignore_fields = NULL)
}
\arguments{
\item{suppress_fields}{tells the formatter to suppress formatting for a specified field name}
}
\value{

}
\description{
Auto categorize format types to pass into 'fmt_acctg_tbl'
}
\examples{
ctgr_acctg_flds(dplyr::rename(mtcars, cyl_cnt = cyl))
}
