% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_acctg_tbl.r
\name{fmt_acctg_tbl}
\alias{fmt_acctg_tbl}
\title{Create an accounting DT format for summary tables}
\usage{
fmt_acctg_tbl(
  data_tbl,
  formats_nms = ctgr_acctg_flds(data_tbl),
  tbl_caption = paste0("Data as of ", Sys.Date()),
  has_totals = T,
  mk_js_buttons = T
)
}
\arguments{
\item{data_tbl}{table of summary accounting values}

\item{formats_nms}{output from function categorize_fields() in format list(percents = c(), dollars = c(), counts = c())}

\item{tbl_caption}{caption to display above table}

\item{has_totals}{whether the acctg table has a totals column}

\item{mk_js_buttons}{render the print and download javascript buttons}
}
\value{
an html table with formatting determined by format_nms
}
\description{
Create an accounting DT format for summary tables
}
\examples{
fmt_acctg_tbl(tibble::tibble(balance = -3:5, percent = seq(.1, .5, by = .05)))
}
